+++ store/types/validity.go
@@ -1,5 +1,7 @@
 package types
 
+import "errors"
+
 var (
 	// MaxKeyLength is the maximum allowed length for a key in bytes.
 	// It is set to 128K - 1 (131,071 bytes).
@@ -25,7 +27,12 @@ func AssertValidValue(value []byte) {
 	if value == nil {
 		panic("value is nil")
 	}
-	if len(value) > MaxValueLength {
-		panic("value is too large")
+	AssertValidValueLength(len(value))
+}
+
+// AssertValidValueLength checks if the value length is within length limit
+func AssertValidValueLength(l int) {
+	if l > MaxValueLength {
+		panic(errors.New("value is too large"))
 	}
 }